rename(#62): api 파일명 변경 mypage → user
